requirements.txt
gymnasium
numpy
opencv-python
pyyaml
